Combine broadcast receivers dealing with notifications into one
