+++ mobile/src/main/AndroidManifest.xml
@@ -277,10 +277,7 @@
             android:name=".background.CopyToClipboardReceiver"
             android:exported="false" />
         <receiver
-            android:name=".core.NotificationDismissedReceiver"
-            android:exported="false" />
-        <receiver
-            android:name=".core.NotificationActionReceiver"
+            android:name=".core.NotificationHandlingReceiver"
             android:exported="false" />
 
         <activity
